Macos build script refactor

+++ rust/build_macos.sh
@@ -1,5 +1,6 @@
 cargo lipo --release --targets aarch64-apple-darwin
 
-rm ../macos/librust_bdk_ffi.dylib
+if [ -e "../macos/librust_bdk_ffi.dylib" ]; then rm ../macos/librust_bdk_ffi.dylib
+        fi
 
 cp target/aarch64-apple-darwin/release/librust_bdk_ffi.dylib ../macos
